😎 (#21) - diary response dto refactor
