--- data/src/main/java/com/example/data/remote/response/GetDiaryListResponse.kt
@@ -1,15 +1,0 @@
-package com.example.data.remote.response
-
-data class GetDiaryListResponse(
-    val diaryList : ArrayList<DiaryResponse>
-){
-    data class DiaryResponse(
-        val id: Int,
-        val title: String,
-        val feeling: String,
-        val image: String,
-        val content: String,
-        val date: String,
-        val writer: String
-        )
-}
